requirements.txt
# Dependencias de Python para MCP Server
fastmcp>=0.1.0
pymongo>=4.0.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
cryptography>=41.0.0  
uvicorn>=0.27.0       

# Para inspeccionar el servidor MCP (TypeScript)
# npx @modelcontextprotocol/inspector python3 src/mongo/MongoMcp.py